requirements.txt
turtle
streamlit
pillow